just merge the code. still need to fix for running
