+++ openrlhf/remote_rm/test.py
@@ -1,0 +1,7 @@
+import requests
+
+url = "http://localhost:1234/predict"
+data = {"query": ["Question: A one-way ticket costs $2. A 30-day pass costs $50. What's the minimum number of rides you will need to take every month so that the 30-day pass is strictly cheaper per ride?\n\nAnswer: Let's think step by step\nIf you take $x$ rides every month, then the cost per ride for the one-way ticket is $\\frac{2x}{x} = 2$ dollars.\nThe cost per ride for the 30-day pass is $\\frac{50}{30} = \\frac{5}{3}$ dollars.\nWe want the 30-day pass to be cheaper per ride, so we need $\\frac{5}{3} < 2$.\nSolving this inequality, we get $5 < 6$, which is always true.\nHowever, we want the 30-day pass to be strictly cheaper per ride, so we need $\\frac{5}{3} < 2$.\nSolving this inequality, we get $5 < 6$, which is always true.\nTherefore, the minimum number of rides you will need to take every month so that the 30-day pass is strictly cheaper per ride is $\\boxed{15}$.\nThe answer is 15", "Question: For real numbers $t,$ the point of intersection of the lines $x + 2y = 7t + 3$ and $x - y = 2t - 2$ is plotted.  All the plotted points lie on a line.  Find the slope of this line.\n\nAnswer: Let's think step by step\nWe can solve the system of equations to find the coordinates of the point of intersection.\nSubtracting the second equation from the first, we get $3y = 5t + 5,$ so $y = \\frac{5t + 5}{3}.$\nSubstituting this into the second equation, we get $x = 2t - 2 + \\frac{5t + 5}{3} = \\frac{11t - 1}{3}.$\nThus, the point of intersection is $\\left(\\frac{11t - 1}{3}, \\frac{5t + 5}{3}\\right).$\nTo find the slope of the line on which all the plotted points lie, we can find the slope of the line passing through any two of these points.\nLet's choose the points corresponding to $t = 0$ and $t = 1.$\nFor $t = 0,$ the point is $\\left(-\\frac{1}{3}, \\frac{5}{3}\\right).$\nFor $t = 1,$ the point is $\\left(\\frac{10}{3}, \\frac{10}{3}\\right).$\nThe slope of the line passing through these points is $\\frac{\\frac{10}{3} - \\frac{5}{3}}{\\frac{10}{3} - \\left(-\\frac{1}{3}\\right)} = \\boxed{\\frac{5}{11}}.$\nThe answer is \\frac{5}{11}", "Question: Luther designs clothes for a high fashion company. His latest line of clothing uses both silk and cashmere fabrics. There are ten pieces made with silk and half that number made with cashmere. If his latest line has thirteen pieces, how many pieces use a blend of cashmere and silk?\n\nAnswer: Let's think step by step\nThere are 10 pieces made with silk.\nHalf that number, which is 5, are made with cashmere.\nSo there are 10 + 5 = 15 pieces made with either silk or cashmere.\nSince the total number of pieces in the line is 13, the number of pieces that use a blend of both fabrics is 15 - 13 = 2.\nThe answer is 2", "Question: Let $\\textrm{A}$ be a digit. If the 7-digit number $353808\\textrm{A}$ is divisible by 2, 3, 4, 5, 6, 8, and 9, then what is $\\textrm{A}$?\n\nAnswer: Let's think step by step\nFor a number to be divisible by 2, its last digit must be even.\nFor a number to be divisible by 5, its last digit must be 0 or 5.\nFor a number to be divisible by 3, the sum of its digits must be divisible by 3.\nFor a number to be divisible by 4, the last two digits must be divisible by 4.\nFor a number to be divisible by 8, the last three digits must be divisible by 8.\nFor a number to be divisible by 9, the sum of its digits must be divisible by 9.\nThe only digit that satisfies all of these conditions is $\\boxed{0}$.\nThe answer is 0"]}
+response = requests.post(url, json=data)
+print(response)
+print(response.json())
